create HeaderBar component
